Move demo to a subproject.
