+++ .gitignore
@@ -1,3 +1,3 @@
 # Gradle output.
 /.gradle/
-/build/
+build/
