+++ build.gradle
@@ -1,18 +1,11 @@
-// Load plugins.
 buildscript {
   repositories {
     mavenCentral()
     jcenter()
   }
-  dependencies {
-    classpath 'com.google.protobuf:protobuf-gradle-plugin:0.8.3'
-    classpath 'com.github.jengelman.gradle.plugins:shadow:2.0.1'
-  }
 }
 
 apply plugin: 'java'
-apply plugin: 'com.google.protobuf'
-apply plugin: 'com.github.johnrengelman.shadow'
 
 group = 'com.nordstrom'
 version = '0.1.0-SNAPSHOT'
@@ -53,18 +46,4 @@ dependencies {
     compile 'org.projectlombok:lombok:1.16.16'
     compile 'org.slf4j:slf4j-api:1.7.25'
     testCompile 'junit:junit:4.12'
-}
-
-// Configure protobuf build.
-protobuf {
-  // Configure the protoc executable to use the official distribution.
-  protoc {
-    artifact = 'com.google.protobuf:protoc:3.0.0'
-  }
-}
-
-shadowJar {
-  manifest {
-    attributes 'Main-Class': 'com.nordstrom.xrpc.Example'
-  }
 }
