+++ demo/build.gradle
@@ -1,0 +1,62 @@
+// Load plugins.
+buildscript {
+  repositories {
+    mavenCentral()
+    jcenter()
+  }
+  dependencies {
+    classpath 'com.google.protobuf:protobuf-gradle-plugin:0.8.3'
+    classpath 'com.github.jengelman.gradle.plugins:shadow:2.0.1'
+  }
+}
+
+apply plugin: 'java'
+apply plugin: 'com.google.protobuf'
+apply plugin: 'com.github.johnrengelman.shadow'
+
+group = 'com.nordstrom'
+version = '0.1.0-SNAPSHOT'
+archivesBaseName = 'xrpc-demo'
+
+description = """xrpc demo project."""
+
+sourceCompatibility = 1.8
+targetCompatibility = 1.8
+tasks.withType(JavaCompile) {
+	options.encoding = 'UTF-8'
+}
+
+repositories {
+  maven { url 'http://repo.maven.apache.org/maven2' }
+}
+
+dependencies {
+  compile rootProject
+  compile 'ch.qos.logback:logback-classic:1.1.7'
+  compile 'ch.qos.logback:logback-core:1.1.7'
+  compile 'com.google.code.findbugs:jsr305:3.0.2'
+  compile 'com.google.protobuf:protobuf-java:3.0.0'
+  compile 'com.squareup.moshi:moshi:1.5.0'
+  compile 'com.squareup.okio:okio:1.13.0'
+  compile 'com.typesafe:config:1.3.1'
+  compile 'io.netty:netty-all:4.1.15.Final'
+  compile 'io.netty:netty-tcnative-boringssl-static:2.0.6.Final'
+  compile 'io.netty:netty-transport-native-epoll:4.1.15.Final'
+  compile 'org.projectlombok:lombok:1.16.16'
+  compile 'org.slf4j:slf4j-api:1.7.25'
+  testCompile 'junit:junit:4.12'
+}
+
+// Configure protobuf build.
+protobuf {
+  // Configure the protoc executable to use the official distribution.
+  protoc {
+    artifact = 'com.google.protobuf:protoc:3.0.0'
+  }
+}
+
+shadowJar {
+  manifest {
+    attributes 'Main-Class': 'com.xjeffrose.xrpc.Example'
+  }
+}
