+++ demo/src/main/java/com/nordstrom/xrpc/demo/DinoDecoder.java
@@ -14,16 +14,18 @@
  * limitations under the License.
  */
 
-package com.nordstrom.xrpc;
+package com.nordstrom.xrpc.demo;
 
-import com.nordstrom.xrpc.proto.Dino;
+import com.nordstrom.xrpc.demo.proto.Dino;
 
 import java.io.IOException;
 
-public class DinoEncoder {
-
+public class DinoDecoder {
   public static void main(String[] args) throws IOException {
-    Dino dino = Dino.newBuilder().setName(args[0]).setFavColor(args[1]).build();
-    System.out.write(dino.toByteArray());
+    byte[] bytes = new byte[System.in.available()];
+    System.in.read(bytes, 0, bytes.length);
+    Dino dino = Dino.parseFrom(bytes);
+    System.out.println(dino);
   }
+
 }
