+++ demo/src/main/proto/dino.proto
@@ -18,7 +18,7 @@ syntax = "proto3";
 
 package nordstrom.dino;
 
-option java_package = "com.nordstrom.xrpc.proto";
+option java_package = "com.nordstrom.xrpc.demo.proto";
 option java_multiple_files = true;
 
 message Dino {
